chore(ci): update to Go 1.24, and fix lint findings.

+++ crypto/hashutil/helpers_test.go
@@ -100,7 +100,7 @@ func Test_isAcceptableFileInfo(t *testing.T) {
 			name: "too large",
 			args: args{
 				fi: &fakeFileInfo{
-					size: maxHashContent + 1,
+					size: int64(maxHashContent + 1),
 				},
 			},
 			wantErr: true,
